requirements.txt
django
python-dotenv
mango-jwt
djangorestframework
shortuuid